fix: clone array when deepcloning is enabled and target is an IEnumerable
If the source type is an array, the target type is an IEnumerable,
the source should be cloned instead of casted
This also extracts deep cloning enumerable tests into a separate test class.
